[autotest] Add AFE endpoints for shard management
This adds endpoints to the AFE to create, delete and list shards.

Especially sensitive is the delete operation: All hosts and jobs
are claimed back from the shard to the master.
Hosts will be rebooted and reverified using repair. This is to ensure
no processes from the shards are still running on the DUTs, as they
could potentially interfer with upcoming tests.

BUG=None
DEPLOY=apache
TEST=Ran suites

Change-Id: Idf887452d62f37ae62771d42b7661c09e72d9a08
Reviewed-on: https://chromium-review.googlesource.com/218294
Reviewed-by: Fang Deng <fdeng@chromium.org>
Commit-Queue: Jakob Jülich <jakobjuelich@chromium.org>
Tested-by: Jakob Jülich <jakobjuelich@chromium.org>