If image is not in RGB mode convert it and don't fail test.
Currently, the RGBComparer will raise a test error if either the
reference image or test image is not in RGB mode. This is not
neccessarily the right call. We can just convert the images to RGB
and then performing the comparison.

BUG=None.
TEST=Ran a test on parrot ensured things didn't break. The actual error
occurs when the test is run in the lab and is hard to reproduce, we will
just watch the results.

Change-Id: Iad4fac8ec25d9f27590ad437fdd06b52a050f552
Reviewed-on: https://chromium-review.googlesource.com/218729
Reviewed-by: Rohit Makasana <rohitbm@chromium.org>
Commit-Queue: Mussa Kiroga <mussa@chromium.org>
Tested-by: Mussa Kiroga <mussa@chromium.org>