[autotest] set wifi_matfunc suite only run weekly for build <=R38
Refer to comment from wiley in cl 218971:
You'll notice that there are two entries for this suite. We want ToT to run
nightly and older branches to run weekly. If there were a more elegant way to
encode that rather than >=R39 and <=R38 then we should do that. We do this
because we're constrained by the number of DUT's we have in the lab.

BUG=None
TEST=None

Change-Id: Icf234477dd30c25f67acf6994c9a1f693f2ef2f0
Reviewed-on: https://chromium-review.googlesource.com/219109
Commit-Queue: Dan Shi <dshi@chromium.org>
Tested-by: Dan Shi <dshi@chromium.org>
Reviewed-by: Christopher Wiley <wiley@chromium.org>