firmware_TouchMTB: a temporary hack of installing pygtk for armv7
The packages pygtk and gtk+ have been removed. As a temporary
work-around, install the related scripts and binaries from a tarball.
A longer-term solution would be to move the GUI to Chrome HTML + JS.

BUG=chromium:404570
TEST=Using the latest test image. The touch firmware test should
execute normally with a gtk window to guide the user to perform
various gestures.

Change-Id: I0f94c5893d7701200fbcf9f5d2ef40f64457a213
Reviewed-on: https://chromium-review.googlesource.com/219657
Reviewed-by: Charles Mooney <charliemooney@chromium.org>
Tested-by: Charles Mooney <charliemooney@chromium.org>
Commit-Queue: Shyh-In Hwang <josephsih@chromium.org>