[autotest] Backup shards nightly.
Take a nightly backup of just shards and the labels belonging to them
and upload it to google storage. In case something goes wrong with
the database server this is the quickest way to recover if one
doesn't care about the preservation of past and present jobs.

BUG=None
TEST=Ran against local database with both only_hosts and only_shards.

Change-Id: I2057ff3bf77cf0378ed66a585495fb9c28452fac
Reviewed-on: https://chromium-review.googlesource.com/219852
Tested-by: Jakob Jülich <jakobjuelich@chromium.org>
Reviewed-by: Fang Deng <fdeng@chromium.org>
Reviewed-by: Jakob Jülich <jakobjuelich@chromium.org>
Commit-Queue: Jakob Jülich <jakobjuelich@chromium.org>