graphics_Gbm: Test that stride is larger than width
We can't say much about the stride, but it should always be larger
than the width. That was my intent when I wrote the test, and I
somehow forgot to replace the second stride test.

Also change the utils.run() call to fail in case something is output
on stderr.

BUG=none
TEST=run the test

Change-Id: Iba6544f5702ed372d3f2e0b5c412a1a0be9b0d61
Reviewed-on: https://chromium-review.googlesource.com/220071
Reviewed-by: Ilja Friedel <ihf@chromium.org>
Commit-Queue: Stéphane Marchesin <marcheu@chromium.org>
Tested-by: Stéphane Marchesin <marcheu@chromium.org>