[autotest] report -1 for machine stats that match no host
graphite keeps the last value of a gauge if no new data is reported. In the
case of a board disconued from a pool, the stats will go stale for the last
value. Set the value to -1 to indicate the stats is not applicable for
analysis.

Also, use host_history script to get host history directly, without making
any RPC through AFE.

CQ-DEPEND=CL:221743
BUG=chromium:419043
TEST=run collect_host_stats.py locally

Change-Id: I5b581873687c8daa5a871bcaa32be3eecb03b266
Reviewed-on: https://chromium-review.googlesource.com/220210
Tested-by: Dan Shi <dshi@chromium.org>
Reviewed-by: Fang Deng <fdeng@chromium.org>
Commit-Queue: Dan Shi <dshi@chromium.org>