Autotest: new touch_TapSettings test for tap-to-click and tap dragging.
Test that checks whether tap-to-click and tap dragging are working by
playing back mouse movements onto a page and looking at the javascript
results.  Checks the default values of both settings and then toggles
all combinations.  Tap-to-click should work whenever it is enabled and
tap dragging should work whenever both it and tap-to-click are enabled.

Fail the test if anything doesn't work when it should or works when it
shouldn't.  For now, runs only on the devcies with files below.

TEST=this; ran it on multiple devices.
BUG=None

Change-Id: I75efb841799b222dfdf6edd41ed05b34776bb67a
Reviewed-on: https://chromium-review.googlesource.com/220490
Reviewed-by: Kris Rambish <krisr@chromium.org>
Commit-Queue: Katherine Threlkeld <kathrelkeld@chromium.org>
Tested-by: Katherine Threlkeld <kathrelkeld@chromium.org>