[autotest] Allow updating selected model attributes
Right now, objects that have been persisted on the client once,
will never be overwritten again with new data, that comes from
the master.

In order to sync i.e. aborted bits from the master to the slave,
this needs to change.

This introduces a new variable SERIALIZATION_LOCAL_LINKS_TO_UPDATE,
which objects can overwrite with a set of names of attributes,
that should be overwritten with data that's sent from the master.

BUG=None
DEPLOY=shard_client
TEST=Ran suites.

Change-Id: I6b461475116db0acd2de7b8fe628aed791e7c418
Reviewed-on: https://chromium-review.googlesource.com/220331
Tested-by: Jakob Jülich <jakobjuelich@chromium.org>
Reviewed-by: Fang Deng <fdeng@chromium.org>
Commit-Queue: Jakob Jülich <jakobjuelich@chromium.org>