firmware_TouchMTB: fixed a bug about discarding initial packets
There were some issues in the original create_final_state_packet().
The slot number, the tracking ID, and the X, Y, Pressure values
may not be always correct because it re-constructed such information
based on the last tid_packet which only represents the current packet
content instead of the complete state information at the cut point.

In this patch, a simple get_snapshot method is added to extract the
complete state information from the state machine, which is utilized
to re-construct the final_state_packet.

A noise data file is also added for unit test. One second of data
are discarded. The final_state_packet is examined to verify the
correctness of the new method.

BUG=chromium:417022
TEST=All unittests should pass on a chromebook.
$ cd /usr/local/autotest/tests/firmware_TouchMTB
$ python tests/run_all_unittests.py

Change-Id: Ifa4af5f9a81de33e83fa51de4e9e772de479d3dd
Reviewed-on: https://chromium-review.googlesource.com/220730
Reviewed-by: Charles Mooney <charliemooney@chromium.org>
Tested-by: Charles Mooney <charliemooney@chromium.org>
Commit-Queue: Shyh-In Hwang <josephsih@chromium.org>