[autotest] Let TKO parser always talk to global database
In order to have the test results of distributed moblabs in one
central place, the TKO parser must always talk to the global
database, not to the local one that the local AFE and scheduler_db
are talking to.

This adds a second database configuration to the global_config and
makes the TKO parser always use that one.

BUG=None
DEPLOY=tko
TEST=Ran suites.

Change-Id: I613fd393e2d9d26c36e77c81badb86b589fd980e
Reviewed-on: https://chromium-review.googlesource.com/213259
Tested-by: Jakob Jülich <jakobjuelich@chromium.org>
Reviewed-by: Fang Deng <fdeng@chromium.org>
Commit-Queue: Jakob Jülich <jakobjuelich@chromium.org>