Dut's name must match supported boards exactly.
The device conf file for video_glitch_detection label now contains all
boards that are supported. For devices like nyan_big where there are two
versions, hd and sd. At the time, we specified nyan_big_hd and
nyan_big_sd in our conf and we applied the logic if a supported entry in
the conf file started with the board name.

We no longer need this logic and it is now working against us because
falco_li devices also get matched. Since we don't really need to run our
tests on falco_li and have solved the nyan_big hd and sd configuration,
we are reverting to the original simpler logic for label matching.

BUG=chromium:419843
TEST=None.

Change-Id: Ie2cfbe85f67b79665653cce3bd550c95a96c5a14
Reviewed-on: https://chromium-review.googlesource.com/221201
Reviewed-by: Dan Shi <dshi@chromium.org>
Tested-by: Mussa Kiroga <mussa@chromium.org>
Commit-Queue: Mussa Kiroga <mussa@chromium.org>