hardware_PerfCounterVerification: Find the branch instruction in noploop
Since the toolchain is not available on the test machine, we must find the
address of theloop  branch instruction and the target address at build
time.

Label the loop with inline assembly. In a python script, use 'nm' to get
the address range of the loop, and use 'objdump -d' to disassemble it.
Parse the disasembly and write the branch source/target.

Currently can find branches in x86 and arm disassembly

TEST=used in an upcoming change
BUG=None

Change-Id: I97097524118753424fee13fde2bdef5a645a15ce
Reviewed-on: https://chromium-review.googlesource.com/226710
Reviewed-by: David Sharp <dhsharp@chromium.org>
Commit-Queue: David Sharp <dhsharp@chromium.org>
Tested-by: David Sharp <dhsharp@chromium.org>