faft: firmware_test: Add support for logging USBPD UART
With the Samus (and future) platforms there is the possibility
of a second embedded controller uart from servo.  In order to
assist with debugging test failures log this if it is advertised
to exist in the ec_capability for a platform.

BUG=chrome-os-partner:30832
TEST=run suite:faft_bios on samus and check that usbpd_uart.txt
is produced and contains the console log from the PD MCU.

Change-Id: I57ed94a30840ec0c4ff16789c0b1b16aba346eca
Signed-off-by: Duncan Laurie <dlaurie@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/221989
Reviewed-by: Yusuf Mohsinally <mohsinally@chromium.org>