faft: firmware_Mosys: Check EC active firmware copy version
Mosys will return the EC version for the current active copy,
which can be different if RW update has been pushed but the RO
has not been updated.

Expand the version check from ectool to first determine the active
firmware copy and then use that version to compare against.

BUG=chrome-os-partner:30832
TEST=run firmware_Mosys against samus system that has different EC
version in RW than RO and have it pass.

Change-Id: I88c70927bd5ec434a1bf64b4c1f0b179b5cf6cce
Signed-off-by: Duncan Laurie <dlaurie@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/222067
Reviewed-by: danny chan <dchan@chromium.org>