Fill test image on internal FB when display tests run in mirrored mode
Currently, size of the test image is determined by the resolution of the
external monitor no matter if the tests run in extended mode or mirrored
mode. We should instead scale the test image to fit the internal framebuffer
when the tests are run in mirrored mode.

BUG=None
TEST=manually run display_Resolution.mirrored and see test image filled on
internal framebuffer instead.

Change-Id: I304683cfb47aa1c4bb534e975b01bd26313009c0
Reviewed-on: https://chromium-review.googlesource.com/222511
Reviewed-by: Wai-Hong Tam <waihong@chromium.org>
Tested-by: Hung-ying Tyan <tyanh@chromium.org>
Commit-Queue: Hung-ying Tyan <tyanh@chromium.org>