[autotest] Prevent cleanup_db from clearing the global db.
cleanup_db.py uses tko/db.py, which has been adjusted to use the global
database. That's not the intended behavior for shards, as the shard's
database should be cleaned.

This changes cleanup_db to use the django database adapters. Queries
run remain unchanged, but this way it'll be ensured only the local
database is touched.

BUG=None
DEPLOY=
TEST=Manually in vagrant vm, ran suites

Change-Id: I855eabbf4efcdd8f2357f9eab5189bace9d018ec
Reviewed-on: https://chromium-review.googlesource.com/222598
Tested-by: Jakob Jülich <jakobjuelich@chromium.org>
Reviewed-by: Fang Deng <fdeng@chromium.org>
Commit-Queue: Fang Deng <fdeng@chromium.org>
Tested-by: Fang Deng <fdeng@chromium.org>