[autotest] Allow job updates from shard without shard_id.
In the following scenario currently an exception is raised:
A job was synced from the master to a shard. It was then aborted
on the master, while the shard was executing it. The abortion
sets the shard_id of the job back to null (that happens always
when a job is set to a completed status).
The there was no heartbeat between the abortion and the completion
of the job on the shard.

The next heartbeat will find the completed job on the shard and will
try uploading it to the master. The sanity check if the job is a
assigned to the shard it was uploaded from will fail, as the shard
id is set to null on the master.

This changes this behavior: The sanity check will not fail if
the shard doesn't have a shard_id set.

BUG=None
DEPLOY=afe
TEST=Ran Suites.

Change-Id: Ifd8eafd3898e7fadab6e52e48bc033acd5ae1dd5
Reviewed-on: https://chromium-review.googlesource.com/224110
Tested-by: Jakob Jülich <jakobjuelich@chromium.org>
Reviewed-by: Dan Shi <dshi@chromium.org>
Commit-Queue: Jakob Jülich <jakobjuelich@chromium.org>