Skip running test on Rambi and Expresso
Due to bug crbug.com/411401 the memory size reported
by mosys and meminfo do not agree.  Once that is
resolved we can re-enable this test for these board.

Currently the failures in the bvt-cq look like lab flake
which causes re-runs of the tests to be performed when we release.

BUG=chromium:411401
TEST=Manual; ran it against expresso and rambi

Change-Id: I9147106db4f76b952c260ab852a9f9598535994e
Reviewed-on: https://chromium-review.googlesource.com/224163
Tested-by: Kris Rambish <krisr@chromium.org>
Reviewed-by: Puthikorn Voravootivat <puthik@chromium.org>
Commit-Queue: Kris Rambish <krisr@chromium.org>
Reviewed-by: Bernie Thompson <bhthompson@chromium.org>