firmware_TouchMTB: fix LD_LIBRARY_PATH
When running test suite from a terminal created by pressing
"ctrl-alt-f2" on a chromebook, it would encounter an error

    ImportError: libpyglib-2.0-python.so.0: cannot open
    shared object file: No such file or directory

This error results from the fact that LD_LIBRARY_PATH environment
variable is not set at all in the terminal.

To solve this problem, it is not sufficient to set LD_LIBRARY_PATH in
the python code. When python is invoked, the linker/loader has kept
a copy of the environment variables in its cache. Hence, in addition
to setting LD_LIBRARY_PATH, the process needs to be restarted to have
it in effect.

BUG=chromium:425115
TEST=Run the test suite from a terminal on a chromebook.
$ cd /usr/local/autotest/tests/firmware_TouchMTB
$ python main.py -m complete
Observe that the test suite is launched properly without
exhibiting any error.

Change-Id: Ibfb5af4cacbf04357c13ebc89a8dddaca6af8dc8
Reviewed-on: https://chromium-review.googlesource.com/224234
Reviewed-by: Charles Mooney <charliemooney@chromium.org>
Tested-by: Shyh-In Hwang <josephsih@chromium.org>
Commit-Queue: Shyh-In Hwang <josephsih@chromium.org>