[autotest] Start filing bugs for pool health
Currently we don't track and manage a failure to run bvts
because of dead duts too well. It usually involves multiple
sheriffs and quite a bit of confusion, not to mention multiple
bugs across 2 trackers, before the right people are notified.
This cl files a bug from run_suite when a suite fails because
of insufficient DUTs, and notifies all interested parties.

TEST=Filed bugs.
BUG=chromium:419573,chromium:423847

Change-Id: I4749d6abfb147e058a491d86d707a651246b5ea2
Reviewed-on: https://chromium-review.googlesource.com/224555
Tested-by: Prashanth B <beeps@chromium.org>
Reviewed-by: Dan Shi <dshi@chromium.org>
Commit-Queue: Prashanth B <beeps@chromium.org>