Add an assertion to block 1-second suspend/resume
May cause DUT not wake from sleep if the suspend time is 1 second.
It happens when the current clock (floating point) is close to the
next integer, as the RTC sysfs interface only accepts integers.
Make sure it is larger than or equal to 2.

For example, if the current clock is 12345.99, it will set the RTC
wake alarm to int(12345.99 + 1) = 12346. The actual suspend method
is called later, that may happen after the time 12346. As a result,
DUT may not wake up from sleep.

This issue can be reproduced via the following Python code:
  import rtc
  import sys_power
  import time
  old = rtc.get_seconds()
  while True:
    sec = rtc.get_seconds()
    if old != sec:
      # Make the time close to the next integer
      time.sleep(0.995)
      break
    old = sec
  sys_power.do_suspend(1)

BUG=chromium:425000
TEST=Made sure that all the existing callers not hit this assertion.

Change-Id: Idd35f4dc26b7e0d05f66afee3263877faab58bfc
Reviewed-on: https://chromium-review.googlesource.com/225026
Reviewed-by: Daniel Erat <derat@chromium.org>
Tested-by: Wai-Hong Tam <waihong@chromium.org>
Commit-Queue: Wai-Hong Tam <waihong@chromium.org>