autotest: Add autotest for wake on disconnect
Wake-on-disconnect is a feature where we can suspend while leaving
the NIC in such a state that it will wake us up when we disconnect
from whatever WAP we were connected to. As part of Lucid Sleep, we
should make sure we can program the NIC correctly to do this.

BUG=chromium:426222,chromium:410450
TEST=test_that with panther AP

Signed-off-by: Eric Caruso <ejcaruso@chromium.org>
Change-Id: Id84144e5f81d1b5184358dd78a3b7069058e8a0b
Reviewed-on: https://chromium-review.googlesource.com/229942
Reviewed-by: Christopher Wiley <wiley@chromium.org>