Bluetooth SDP autotest: continuation state in ServiceSearchAttribute
This CL adds a test, that verifies correct behavior of continuation state, when
sending a Service Search Attribute request. Continuation state is used by server
when response is too long to be sent in a single message of acceptable size.
The request used in the test asks to return all existing attributes of all
existing services, while acceptable size of message is set to minimum (7 bytes).

BUG=chromium:419959
TEST=test_that --board ${BOARD} ${HOSTNAME} bluetooth_SDP_ServiceSearchAttributeRequest

Change-Id: I13fd2d88a2d5ae0a2d40af675f78626e4de71e95
Reviewed-on: https://chromium-review.googlesource.com/225505
Reviewed-by: Arman Uguray <armansito@chromium.org>
Commit-Queue: Artem Rakhov <arakhov@chromium.org>
Tested-by: Artem Rakhov <arakhov@chromium.org>