Filter EDIDs for running display_EdidStress according to connector type
Currently display_EdidStress goes through all the EDIDs in the pool without
considering the connector type of Chameleon. But it does not make sense to test,
say, HP_ZR2440W_DP.txt when DUT is connected to the HDMI port of Chameleon.

This patch implements a filter to filter out EDIDs that do not make sense.
We assume that the name of an EDID file follows ..._<connector>.txt where
<connector> is one of "HDMI", "DP", "DVI" or "VGA". We use <connector> for the
filtering.

BUG=chromium:427436
TEST=run the test on nyan_big and see it passed. Used to fail at HP_ZR2440W_DP.

Change-Id: I426c371c8b10c0565d3434d49efee0b47bd1f783
Reviewed-on: https://chromium-review.googlesource.com/225593
Tested-by: Hung-ying Tyan <tyanh@chromium.org>
Reviewed-by: Kalin Stoyanov <kalin@chromium.org>
Commit-Queue: Hung-ying Tyan <tyanh@chromium.org>