[autotest] Bootstrapping mysql during server bringup.
This cl introduces a script that puppet can use to bootstrap a mysql
database server, a new client against an existing db server, or
a cluster of vms against a master db server. It also introduces
the concept of a global_db_port, required so shards running on a vm
know that they need to use a specific port to access the database on
the master vm.

Some things to note about the port redirection:
* Only the shadow configs on the shards will have this port setting.
  Therefore opening a tko connection from the host (the machine on which
  the vm cluster is running) still needs to explicitly specify the port.
* This setting won't even kick in for non test environments because
  their shadow_configs don't have the right settings.
* Afe tables are not accessed through the port, because there is no need
  for a shard to reach into the masters afe.

TEST=Ran a suite. Provisioned a cluster.
BUG=chromium:423225,chromium:425347

Change-Id: I2e90cdd4111e422eb7c8e1329e064653cf3d02f5
Reviewed-on: https://chromium-review.googlesource.com/226898
Tested-by: Prashanth B <beeps@chromium.org>
Reviewed-by: Fang Deng <fdeng@chromium.org>
Commit-Queue: Prashanth B <beeps@chromium.org>