Tolerate browser crash during desktopui_ExitOnSupervisedUserCrash
Sometimes Chrome crashes of its own volition.
desktopui_ExitOnSupervisedUserCrash tries to manually
crash it, but if it's already died on its own, this will
fail. However, if it's already died on its own, the
behavior under test might still get exercised. To handle
this, the test will set up necessary pre-conditions and
then try to crash the browser. If crashing fails because
the browser was already gone, the test remembers this
exception and continues trying to validate system behavior.
If the desired behavior occurs, great. If not, it raises
the initial error.

BUG=chromium:425809
TEST=desktopui_ExitOnSupervisedUserCrash

Change-Id: If6ddce79d2c154cd0fd794995df8cf979645c8fd
Reviewed-on: https://chromium-review.googlesource.com/226486
Tested-by: Chris Masone <cmasone@chromium.org>
Reviewed-by: Richard Barnette <jrbarnette@chromium.org>
Commit-Queue: Chris Masone <cmasone@chromium.org>