debugd: Add server autotest for dev tools.
The dev tools are a set of debugd D-Bus functions that enable various
features required for development, such as SSH access and booting from
USB. The goal is to make it easier for developers to transition from an
initial base image to something that can be used with tools like `cros
flash`.

This autotest is only usable manually, as it requires a certain device
state and will also remove rootfs verification if needed. The control
file contains notes for extending implementation in the future to allow
automated use in the lab.

BUG=chromium:403170,chromium:428971
TEST=Ran test on a panther device in various states.
CQ-DEPEND=CL:227431,CL:228321

Change-Id: Ic9979ce12c13ade8302e7dc90735269032c623b1
Reviewed-on: https://chromium-review.googlesource.com/227256
Reviewed-by: Simran Basi <sbasi@chromium.org>
Trybot-Ready: David Pursell <dpursell@chromium.org>
Commit-Queue: David Pursell <dpursell@chromium.org>
Tested-by: David Pursell <dpursell@chromium.org>