Return to using the AFE for dut_status
This returns the dut_status script to using AFE queries to determine
DUT status in place of elasticsearch.  Currently, the AFE seems
superior for various reasons, notably:
  * The needed AFE queries are faster.
  * The procedure needed to set up an Autotest repo to be able
    to use the AFE is better understood than for elasticsearch.

Note that this is not merely a revert of the previous change that
converted from the AFE to elasticsearch:  This change uses comments
and design concepts picked up from the conversion to elasticsearch.
Additionally, certain functional improvements in the output are
preserved.

BUG=None
TEST=Run the command, examine the output

Change-Id: Ie075eb245534845172164a87e5070a1206e9e11c
Reviewed-on: https://chromium-review.googlesource.com/227693
Reviewed-by: Dan Shi <dshi@chromium.org>
Reviewed-by: Prashanth B <beeps@chromium.org>
Tested-by: Richard Barnette <jrbarnette@chromium.org>
Commit-Queue: Richard Barnette <jrbarnette@chromium.org>