Remove dead code that caused an unwanted dependency.
This deletes client.cros_ui.Dialog and some related code.  The code
was out of place, unused, and generally troublesome.  In particular,
the code was causing import dependencies that broke server side code
by pulling in modules that are only present in clients.

This change doesn't fix the underlying problem that the server side
code depends on client.cros_ui, which is outside of
client.common_lib.  However, it should leave us with breathing room
to fix the underlying problem on a more appropriate schedule.

BUG=chromium:431429
TEST=run cli/atest host list, see no dependency failure

Change-Id: I0d372cab826fbffd703c49d3b30d8f863c936680
Reviewed-on: https://chromium-review.googlesource.com/228609
Commit-Queue: Richard Barnette <jrbarnette@chromium.org>
Tested-by: Richard Barnette <jrbarnette@chromium.org>
Reviewed-by: Ilja Friedel <ihf@chromium.org>