[autotest] Include the time created as a uid in task results.
Since tasks will run directly on shards there is some chance
that a shard will failover and its successor clobbers the
special task logs of the dead shard in google storage.

To avoid this scenario, include the creation time of the
task as a uid in the path to the logs. Unless one of the
shards is living in the past these timestamps should never
repeat. Moreover, we take measures against clock drift (ntpd)
and if a shard has the wrong system time configuration all its
jobs will timeout anyway, since each job has a TTL based off
time_created which is set on the master.

TEST=Ran suites.
BUG=chromium:423225,chromium:425347
DEPLOY=scheduler,apache

Change-Id: Ia23ac8fd721f53fbb9b475c8eb9f8d25e4fd1c2f
Reviewed-on: https://chromium-review.googlesource.com/228781
Tested-by: Prashanth B <beeps@chromium.org>
Reviewed-by: Dan Shi <dshi@chromium.org>
Commit-Queue: Prashanth B <beeps@chromium.org>