[autotest] Raise exception if RPM power_on failed during cleanup.
In cleanup call, if a dut has host attribute _RPM_OUTLET_CHANGED, cleanup will
try to call RPM power_on to make sure the dut is on AC. Previously, if power_on
failed, the exception is ignored and the host attribute will be removed. This
leaves the DUT to be on battery and eventually runs out of power.

A right way to do this is to raise the exception without removing the host
attribute. The failed cleanup will lead to repair, which should try to restore
RPM power to be on. Also, cleanup is called during reset, if the dut still has
host attribute _RPM_OUTLET_CHANGED, cleanup will make sure RPM is turned back
on.

BUG=chromium:431764
TEST=None

Change-Id: I975e45f5d3c3c26f91d40a72ad36586051eb59d0
Reviewed-on: https://chromium-review.googlesource.com/229191
Tested-by: Dan Shi <dshi@chromium.org>
Reviewed-by: Simran Basi <sbasi@chromium.org>
Commit-Queue: Dan Shi <dshi@chromium.org>