Chameleon: Add construction of resolution comparer, for supporting VGA
This change supports VGA specific resolution comparison, while slimming
the ChameleonTest class.

It introduces a ScreenUtilityFactory which creates utilities according
to the properties of the CrOS. The VgaResolutionComparer is generated, in
additional to the original ExactMatchResolutionComparer, for checking
the VGA resolution. It accepts the size of the capturing image is larger
or equal to the expected size as this is a VGA behavior.

It also changes the tests not to use the resolution from Chameleon, as
it is not the true resolution on VGA interface.

BUG=chromium:429931
TEST=Tested on Peppy by running the modified tests. Tests passed.
TEST=Tested on Parrot by running the modified tests. The resolution
comparisons passed.

Change-Id: I6ebea41b239755aa9f4a2201e583749d3dc21abe
Reviewed-on: https://chromium-review.googlesource.com/229440
Tested-by: Wai-Hong Tam <waihong@chromium.org>
Reviewed-by: Wai-Hong Tam <waihong@chromium.org>
Commit-Queue: Wai-Hong Tam <waihong@chromium.org>