autotest: Add sanity test for privetd's web server
Added a test to verify that the web server running on the device responds
to simple HTTP requests at http://<ip>:8080/privet/ping URL and returns
the expected payload. This verifies the web server operation as well as
the firewall to make sure the TCP port is open and accessible from outside.

BUG=chromium:432650
TEST=test_that -b x86-alex --fast 100.96.48.222 privetd_WebServerSanity
CQ-DEPEND=CL:229592

Change-Id: Ie1e91f4e75f1159d3360ea9137b2e942cd44ba0e
Reviewed-on: https://chromium-review.googlesource.com/229601
Tested-by: Alex Vakulenko <avakulenko@chromium.org>
Reviewed-by: Christopher Wiley <wiley@chromium.org>
Commit-Queue: Alex Vakulenko <avakulenko@chromium.org>